+++ src/api/index.js
@@ -1,111 +1,3 @@
-import {EventsModel} from "../model/index";
-
-const HttpMethod = {
-  GET: `GET`,
-  PUT: `PUT`,
-  POST: `POST`,
-  DELETE: `DELETE`
-};
-
-const SuccessHTTPStatusRange = {
-  MIN: 200,
-  MAX: 299
-};
-
-export default class Api {
-  constructor(endPoint, authorization) {
-    this._endPoint = endPoint;
-    this._authorization = authorization;
-  }
-
-  getEvents() {
-    return this._load({url: `points`})
-      .then(Api.toJSON)
-      .then((points) => points.map(EventsModel.adaptToClient));
-  }
-
-  getDestinations() {
-    return this._load({url: `destinations`})
-      .then(Api.toJSON);
-  }
-
-  getOffers() {
-    return this._load({url: `offers`})
-      .then(Api.toJSON);
-  }
-
-  updateEvent(event) {
-    return this._load({
-      url: `points/${event.id}`,
-      method: HttpMethod.PUT,
-      body: JSON.stringify(EventsModel.adaptToServer(event)),
-      headers: new Headers({"Content-Type": `application/json`})
-    })
-      .then(Api.toJSON)
-      .then(EventsModel.adaptToClient);
-  }
-
-  addEvent(event) {
-    return this._load({
-      url: `points`,
-      method: HttpMethod.POST,
-      body: JSON.stringify(EventsModel.adaptToServer(event)),
-      headers: new Headers({"Content-Type": `application/json`})
-    })
-      .then(Api.toJSON)
-      .then(EventsModel.adaptToClient);
-  }
-
-  deleteEvent(event) {
-    return this._load({
-      url: `points/${event.id}`,
-      method: HttpMethod.DELETE
-    });
-  }
-
-  sync(data) {
-    return this._load({
-      url: `points/sync`,
-      method: HttpMethod.POST,
-      body: JSON.stringify(data),
-      headers: new Headers({"Content-Type": `application/json`})
-    })
-      .then(Api.toJSON);
-  }
-
-
-  _load({
-    url,
-    method = HttpMethod.GET,
-    body = null,
-    headers = new Headers()
-  }) {
-    headers.append(`Authorization`, this._authorization);
-
-    return fetch(
-        `${this._endPoint}/${url}`,
-        {method, body, headers}
-    )
-      .then(Api.checkStatus)
-      .catch(Api.catchError);
-  }
-
-  static checkStatus(response) {
-    if (
-      response.status < SuccessHTTPStatusRange.MIN &&
-      response.status > SuccessHTTPStatusRange.MAX
-    ) {
-      throw new Error(`${response.status}: ${response.statusText}`);
-    }
-
-    return response;
-  }
-
-  static toJSON(response) {
-    return response.json();
-  }
-
-  static catchError(err) {
-    throw err;
-  }
-}
+export {default as Api} from "./api";
+export {default as Store} from "./store";
+export {default as Provider} from "./provider";
